refactor(notes): split HTML site export modules
